demo/graph_tutorial/requirements.txt
asgiref==3.2.7
certifi==2019.11.28
chardet==3.0.4
Django==2.2.14
idna==2.9
oauthlib==3.1.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
